06-pex-ctx-00 Context cleanup initial stack state and length check on pop
This is related to @14eab59bb2c9acc09bd306db0605715e5791811d. We don’t
need stack states filled initially. At the beginning we compared the
previously pushed state with the previous state -1, which was wrong.
Now we compare the current set states with the previously pushed ones.
We alter the current state variables, push them, and on pop take the
pushed ones again and compare them with the state variables set. So we
only need the top of the stack.